backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.2
pydantic-settings>=2.1.0
alembic>=1.12.1
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
pypdf>=3.17.1
pandas>=2.1.0
scikit-learn>=1.3.0
numpy>=1.26.0
joblib>=1.3.2
python-dotenv>=1.0.0
httpx>=0.25.2
pytest>=7.4.3
tenacity>=8.2.3
jinja2
pyinstaller
pywebview
google-generativeai